preparing for products_attributes_iframe

+++ admin/includes/modules/new_product.php
@@ -233,6 +233,9 @@
         if (isset($_GET['pID']) && $_GET['pID'] > 0) {
 					if (is_file('includes/modules/products_attributes_iframe.php')) {
 					  include_once("includes/modules/products_attributes_iframe.php");
+					}
+          if (function_exists('attributes_iframe_link')) {
+						echo '&nbsp;&nbsp;'.attributes_iframe_link($_GET['pID']);
 					} else {
             echo '&nbsp;&nbsp;<a class="button" href="' . xtc_href_link('new_attributes.php','cpath='. $cPath . $catfunc->page_parameter.'&current_product_id='.$_GET['pID'].'&action=edit&oldaction=new_product').'" onclick="this.blur()">'.BUTTON_EDIT_ATTRIBUTES.'</a>';
           }
